pyproject.toml
[tool.poetry]
name = "ig-markets-api"
version = "0.1.0"
description = ""
authors = ["Kevin Courbet <kevin.courbet.contact@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flake8 = "^3.8.4"
autopep8 = "^1.5.4"
attrs = "^20.3.0"
httpx = "^0.16.1"
pydantic = {extras = ["email", "dotenv"], version = "^1.7.2"}
loguru = "^0.5.3"
tenacity = "^6.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
ipykernel = "^5.3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
